+++ src/main/groovy/com/monits/gradle/sca/ToolVersions.groovy
@@ -32,7 +32,7 @@ final class ToolVersions {
     private final static GradleVersion GRADLE_VERSION_CHECKSTYLE = GradleVersion.version('2.7')
 
     private final static String FINDBUGS_TOOL_VERSION = '3.0.1'
-    private final static String FINDBUGS_MONITS_VERSION = '0.2.0-SNAPSHOT'
+    private final static String FINDBUGS_MONITS_VERSION = '0.2.0'
     private final static String FB_CONTRIB_VERSION = '6.6.3'
 
     private final static String TOOL_GRADLE = 'Gradle'